Create a structure for the application
A lot of work is done here to put a solid struction on the
application.

This application structure is  adopted because it allows the
application to be deploable as both monolithic or micro-service.
It also allows the various microservices to share code logic (pkg)
without being tightly coupled together.

The code logic is grouped by domains (/internals), and each domain
can can be deployed via their respective server
(e.g. /internal/<domain>/server/http.go)